worknest/src/app/backend/micro_event/requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.6
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0